feat: lbac 1903: amelioration de la page d'administration d'une entreprise
